+++ packages/wallet-ui/src/components/ui/organism/Header/SendModal/SendModal.view.tsx
@@ -5,28 +5,8 @@ import { useAppSelector, useCurrentNetwork, useEstimateFee } from 'hooks';
 import { FeeToken } from 'types';
 import { useMultiLanguage, useStarkNetSnap } from 'services';
 import { DEFAULT_FEE_TOKEN } from 'utils/constants';
-import {
-  getMaxAmountToSpend,
-  isValidAddress,
-  isValidStarkName,
-  shortenAddress,
-} from 'utils/utils';
-import { AmountInput } from 'components/ui/molecule/AmountInput';
-import { AddressInput } from 'components/ui/molecule/AddressInput';
-import { DropDown } from 'components/ui/molecule/DropDown';
-import {
-  Bold,
-  Normal,
-} from 'components/ui/organism/ConnectInfoModal/ConnectInfoModal.style';
-import { InfoText } from 'components/ui/molecule/AddressInput/AddressInput.style';
-import {
-  ButtonStyled,
-  MessageAlert,
-  Network,
-  Separator,
-  SeparatorSmall,
-} from './SendModal.style';
-import { Modal } from 'components/ui/atom/Modal';
+import { isValidAddress, isValidStarkName } from 'utils/utils';
+import { SendInputModal } from '../SendInputModal';
 import { SendSummaryModal } from '../SendSummaryModal';
 
 interface Props {
@@ -34,7 +14,6 @@ interface Props {
 }
 
 export const SendModalView = ({ closeModal }: Props) => {
-  const networks = useAppSelector((state) => state.networks);
   const chainId = useCurrentNetwork()?.chainId;
   const erc20TokenBalanceSelected = useAppSelector(
     (state) => state.wallet.erc20TokenBalanceSelected,
@@ -154,79 +133,19 @@ export const SendModalView = ({ closeModal }: Props) => {
   return (
     <>
       {!summaryModalOpen && (
-        <Modal>
-          <Modal.Title>{translate('send')}</Modal.Title>
-          <Modal.Body>
-            <Network>
-              <Normal>{translate('network')}</Normal>
-              <Bold>{networks.items[networks.activeNetwork].name}</Bold>
-            </Network>
-            <AddressInput
-              label={translate('to')}
-              placeholder={translate('pasteRecipientAddress')}
-              onChange={(value) => handleChange('address', value.target.value)}
-              value={fields.address}
-              disableValidate
-              validateError={errors.address}
-            />
-            {isValidStarkName(fields.address) && resolvedAddress && (
-              <InfoText>{shortenAddress(resolvedAddress, 12)}</InfoText>
-            )}
-            <SeparatorSmall />
-            <MessageAlert
-              variant="info"
-              text={translate('validStarknetAddressOnly')}
-            />
-            <Separator />
-            <AmountInput
-              label={translate('amount')}
-              onChangeCustom={(value) => handleChange('amount', value)}
-              value={fields.amount}
-              error={errors.amount !== '' ? true : false}
-              helperText={errors.amount}
-              decimalsMax={erc20TokenBalanceSelected.decimals}
-              asset={{
-                ...erc20TokenBalanceSelected,
-                amount: getMaxAmountToSpend(
-                  erc20TokenBalanceSelected,
-                  feeEstimates?.fee,
-                ),
-              }}
-              isFetchingFee={loading}
-              setShouldApplyMax={setShouldApplyMax}
-              shouldApplyMax={shouldApplyMax}
-            />
-            <SeparatorSmall />
-            <div>
-              <label htmlFor="feeToken">
-                {translate('selectTokenForTransactionFees')}
-              </label>
-              <DropDown
-                value={fields.feeToken}
-                options={Object.values(FeeToken).map((token) => ({
-                  label: token,
-                  value: token,
-                }))}
-                onChange={(e) => handleChange('feeToken', e.value)}
-              />
-            </div>
-          </Modal.Body>
-          <Modal.Buttons>
-            <ButtonStyled
-              onClick={closeModal}
-              backgroundTransparent
-              borderVisible
-            >
-              {translate('cancel')}
-            </ButtonStyled>
-            <ButtonStyled
-              onClick={() => setSummaryModalOpen(true)}
-              enabled={confirmEnabled()}
-            >
-              {translate('confirm')}
-            </ButtonStyled>
-          </Modal.Buttons>
-        </Modal>
+        <SendInputModal
+          closeModal={closeModal}
+          setSummaryModalOpen={setSummaryModalOpen}
+          confirmEnabled={confirmEnabled}
+          handleChange={handleChange}
+          fields={fields}
+          errors={errors}
+          resolvedAddress={resolvedAddress}
+          loading={loading}
+          shouldApplyMax={shouldApplyMax}
+          setShouldApplyMax={setShouldApplyMax}
+          feeEstimates={feeEstimates}
+        />
       )}
       {summaryModalOpen && (
         <SendSummaryModal
